requirements.txt
openai
pydantic
python-dotenv
PyMuPDF # For reading PDFs

fastapi
uvicorn